Add ship initialization and fixed thrust calculation
Added an `InitializeShip` method to the `ShipContainer` class, which performs initial calculations on the ship object. Made changes to the thrust calculation to be based on the thruster block's direction rather than its position. Reduced window dimensions and simplified expressions across the `default-2022.dwlt` file for consistency and simplicity reasons.

Signed-off-by: Icewav3 <coledorman12@gmail.com>