Refactor turret controllers and delete unused prefab
Refactored code for turret controllers splitting it into Laser and Kinetic turret controller classes. This involved moving TurretController.cs to a new directory and renaming it to LaserTurretController.cs. The KineticTurretController.cs is a newly added file. Also included in this commit is the deletion of an unused TurretBlockBase.prefab file for effective code management and improved performance.